move xs tests
